requirements.txt
ImageHash~=4.3.1
pillow~=10.3.0
beautifulsoup4~=4.13.0b2